Add files via upload
I apologize for forgetting to create the repo before I started the project.
In the slack channel I sent a screenshot of what was happening when I tried to run the debugger, it was not working so I had to answer all the questions without knowing if I got any of them right. All I could do was write the code.